requirements.txt
langchain>=0.2.10,<0.3.0
langchain-openai>=0.1.7,<0.2.0
langchain-community>=0.2.7,<0.3.0
faiss-cpu>=1.7.4,<2.0.0
openai>=1.33.0,<2.0.0
streamlit>=1.33.0,<2.0.0
yfinance>=0.2.36,<0.3.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.31.0,<3.0.0
tavily-python>=0.3.3,<0.4.0
beautifulsoup4>=4.12,<5.0
lxml>=5.2,<6.0
pandas>=2.1,<3.0
numpy>=1.26,<3.0
requests>=2.31,<3.0


